# backend-go/openapi.yaml
openapi: 3.0.3
info:
  title: Cancer Detection Backend API
  version: 1.0.0
  description: |
    Gateway-facing backend API.
    Internal microservice communication may use gRPC, but external clients should use these HTTP endpoints.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/reports:
    get:
      tags: [Reports]
      summary: List reports
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: patientId
          schema: { type: string }
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items: { type: object, additionalProperties: true }
                  nextCursor: { type: string, nullable: true }
    post:
      tags: [Reports]
      summary: Create or upsert report (DOCTOR/ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/reports/{id}:
    get:
      tags: [Reports]
      summary: Get report by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Reports]
      summary: Update report (DOCTOR/ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: string }
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
...
    ReportCreateRequest:
      type: object
      required: [detectionId, patientId]
      properties:
        detectionId: { type: string }
        patientId: { type: string }
        content:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [DRAFT, FINALIZED]
          default: DRAFT
    ReportUpdateRequest:
      type: object
      properties:
        content:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [DRAFT, FINALIZED]
        pdfPath:
          type: string
...